+++ ch11/edgeimpulse_ex/edge-impulse-sdk/CMSIS/DSP/Source/SupportFunctions/arm_barycenter_f16.c
@@ -5,11 +5,13 @@
  * Title:        arm_barycenter_f16.c
  * Description:  Barycenter
  *
+ * $Date:        23 April 2021
+ * $Revision:    V1.9.0
  *
  * Target Processor: Cortex-M and Cortex-A cores
  * -------------------------------------------------------------------- */
 /*
- * Copyright (C) 2010-2020 ARM Limited or its affiliates. All rights reserved.
+ * Copyright (C) 2010-2021 ARM Limited or its affiliates. All rights reserved.
  *
  * SPDX-License-Identifier: Apache-2.0
  *
@@ -105,7 +107,7 @@ void arm_barycenter_f16(const float16_t *in,
         w2 = *pW++;
         w3 = *pW++;
         w4 = *pW++;
-        accum += w1 + w2 + w3 + w4;
+        accum += (_Float16)w1 + (_Float16)w2 + (_Float16)w3 + (_Float16)w4;
 
         blkCntSample = vecDim >> 3;
         while (blkCntSample > 0) {
@@ -131,10 +133,10 @@ void arm_barycenter_f16(const float16_t *in,
 
         blkCntSample = vecDim & 7;
         while (blkCntSample > 0) {
-            *pOut = *pOut + *pIn1++ * w1;
-            *pOut = *pOut + *pIn2++ * w2;
-            *pOut = *pOut + *pIn3++ * w3;
-            *pOut = *pOut + *pIn4++ * w4;
+            *pOut = (_Float16)*pOut + (_Float16)*pIn1++ * (_Float16)w1;
+            *pOut = (_Float16)*pOut + (_Float16)*pIn2++ * (_Float16)w2;
+            *pOut = (_Float16)*pOut + (_Float16)*pIn3++ * (_Float16)w3;
+            *pOut = (_Float16)*pOut + (_Float16)*pIn4++ * (_Float16)w4;
             pOut++;
             blkCntSample--;
         }
@@ -156,7 +158,7 @@ void arm_barycenter_f16(const float16_t *in,
 
         pOut = out;
         w = *pW++;
-        accum += w;
+        accum += (_Float16)w;
 
         blkCntSample = vecDim >> 3;
         while (blkCntSample > 0) 
@@ -174,7 +176,7 @@ void arm_barycenter_f16(const float16_t *in,
         blkCntSample = vecDim & 7;
         while (blkCntSample > 0) 
         {
-            *pOut = *pOut + *pIn++ * w;
+            *pOut = (_Float16)*pOut + (_Float16)*pIn++ * (_Float16)w;
             pOut++;
             blkCntSample--;
         }
@@ -184,7 +186,7 @@ void arm_barycenter_f16(const float16_t *in,
 
     /* Normalize */
     pOut = out;
-    accum = 1.0f / accum;
+    accum = 1.0f16 / (_Float16)accum;
 
     blkCntSample = vecDim >> 3;
     while (blkCntSample > 0) 
@@ -201,7 +203,7 @@ void arm_barycenter_f16(const float16_t *in,
     blkCntSample = vecDim & 7;
     while (blkCntSample > 0) 
     {
-        *pOut = *pOut * accum;
+        *pOut = (_Float16)*pOut * (_Float16)accum;
         pOut++;
         blkCntSample--;
     }
@@ -218,7 +220,7 @@ void arm_barycenter_f16(const float16_t *in, const float16_t *weights, float16_t
    blkCntVector = nbVectors;
    blkCntSample = vecDim;
 
-   accum = 0.0f;
+   accum = 0.0f16;
 
    pW = weights;
    pIn = in;
@@ -229,7 +231,7 @@ void arm_barycenter_f16(const float16_t *in, const float16_t *weights, float16_t
 
    while(blkCntSample > 0)
    {
-         *pOut = 0.0f;
+         *pOut = 0.0f16;
          pOut++;
          blkCntSample--;
    }
@@ -239,12 +241,12 @@ void arm_barycenter_f16(const float16_t *in, const float16_t *weights, float16_t
    {
       pOut = out;
       w = *pW++;
-      accum += w;
+      accum += (_Float16)w;
 
       blkCntSample = vecDim;
       while(blkCntSample > 0)
       {
-          *pOut = *pOut + *pIn++ * w;
+          *pOut = (_Float16)*pOut + (_Float16)*pIn++ * (_Float16)w;
           pOut++;
           blkCntSample--;
       }
@@ -258,7 +260,7 @@ void arm_barycenter_f16(const float16_t *in, const float16_t *weights, float16_t
 
    while(blkCntSample > 0)
    {
-         *pOut = *pOut / accum;
+         *pOut = (_Float16)*pOut / (_Float16)accum;
          pOut++;
          blkCntSample--;
    }
